# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Plexus API
  description: Enterprise project management API
  version: 1.0.0
  contact:
    name: Plexus
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://api.plexus.app/api/v1
    description: Production
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        logo_url:
          type: string
          nullable: true
        plan:
          type: string
          enum: [free, pro, enterprise]
        my_role:
          type: string
          enum: [owner, admin, member, guest]
      required: [id, slug, name, plan]
...
    OrgMember:
      type: object
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
        email:
          type: string
        avatar_url:
          type: string
          nullable: true
        role:
          type: string
          enum: [owner, admin, member, guest]
        joined_at:
          type: string
          format: date-time
      required: [id, display_name, email, role, joined_at]
...
  parameters:
    orgSlug:
      name: orgSlug
      in: path
      required: true
      schema:
        type: string
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /auth/oidc/login:
    get:
      tags: [Auth]
      operationId: oidcLogin
      security: []
      description: Redirects to the configured OIDC provider authorization endpoint.
      responses:
        '302':
          description: Redirect to OIDC provider
        '501':
          description: OIDC is not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/oidc/callback:
    get:
      tags: [Auth]
      operationId: oidcCallback
      security: []
      description: OIDC callback; provisions user on first login and redirects to the frontend with tokens.
      parameters:
        - name: code
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: error
          in: query
          schema:
            type: string
      responses:
        '302':
          description: Redirect to FRONTEND_URL/auth/callback with access_token and refresh_token query params
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '501':
          description: OIDC is not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /orgs:
    post:
      tags: [Organizations]
      operationId: createOrg
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                slug:
                  type: string
              required: [name]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
    get:
      tags: [Organizations]
      operationId: listMyOrgs
      responses:
        '200':
          description: List of organizations
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Organization'
...
  /orgs/{orgSlug}:
    get:
      tags: [Organizations]
      operationId: getOrg
      parameters:
        - $ref: '#/components/parameters/orgSlug'
      responses:
        '200':
          description: Organization details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
    patch:
      tags: [Organizations]
      operationId: updateOrg
      parameters:
        - $ref: '#/components/parameters/orgSlug'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                logo_url:
                  type: string
      responses:
        '204':
          description: Updated
  /orgs/{orgSlug}/members:
    get:
      tags: [Organizations]
      operationId: listOrgMembers
      parameters:
        - $ref: '#/components/parameters/orgSlug'
      responses:
        '200':
          description: Org members
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrgMember'
  /orgs/{orgSlug}/members/invite:
    post:
      tags: [Organizations]
      operationId: inviteMember
      parameters:
        - $ref: '#/components/parameters/orgSlug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                role:
                  type: string
                  enum: [owner, admin, member, guest]
              required: [email]
      responses:
        '204':
          description: Invited
  /orgs/{orgSlug}/members/{userID}:
    delete:
      tags: [Organizations]
      operationId: removeMember
      parameters:
        - $ref: '#/components/parameters/orgSlug'
        - name: userID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Removed
...